+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Breaking changes
 
 * Rename `DefaultTrap` to `BestEffortTrap` for better clarity.
+* Rename `logforth_core::record::Metadata` to `FilterCriteria`.
 * Redesign `LevelFilter` to allow different comparison methods.
 * Redesign `Level` as opentelemetry severity numbers.
     * Add `Level::Fatal` variant to represent fatal level logs.